refacotor recognized_major
