+++ snugh/lecture/deprecated.py
@@ -1,0 +1,148 @@
+# Deprecated Common Functions
+def update_plan_info(plan):
+    # TODO: need to delete
+    # SemesterLecture 모델의 lecture_type 관련 값 업데이트
+    majors = Major.objects.filter(planmajor__plan=plan)
+    entrance_year = plan.user.userprofile.entrance_year
+    semester_lectures = SemesterLecture.objects.filter(semester__plan=plan)
+    for semester_lecture in list(semester_lectures):
+        lecture = semester_lecture.lecture
+        recognized_major, lecture_type = cal_lecture_type(lecture, majors, entrance_year)
+        semester_lecture.recognized_major = recognized_major
+        semester_lecture.lecture_type = lecture_type
+        semester_lecture.save()
+
+    # Semester 모델의 credit 관련 값 업데이트
+    semesters = Semester.objects.filter(plan=plan)
+    for semester in list(semesters):
+        mr_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                             semesterlecture__lecture_type=SemesterLecture.MAJOR_REQUIREMENT)
+        me_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                             semesterlecture__lecture_type=SemesterLecture.MAJOR_ELECTIVE)
+        t_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                            semesterlecture__lecture_type=SemesterLecture.TEACHING)
+        g_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                            semesterlecture__lecture_type=SemesterLecture.GENERAL)
+        ge_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                             semesterlecture__lecture_type=SemesterLecture.GENERAL_ELECTIVE)
+        mr_credit = 0
+        me_credit = 0
+        g_credit = 0
+        ge_credit = 0
+
+        for lecture in list(mr_lectures):
+            mr_credit += lecture.credit
+        for lecture in list(me_lectures):
+            me_credit += lecture.credit
+        for lecture in list(t_lectures):
+            me_credit += lecture.credit
+        for lecture in list(g_lectures):
+            g_credit += lecture.credit
+        for lecture in list(ge_lectures):
+            ge_credit += lecture.credit
+
+        semester.major_requirement_credit = mr_credit
+        semester.major_elective_credit = me_credit
+        semester.general_credit = g_credit
+        semester.general_elective_credit = ge_credit
+        semester.save()
+
+    # PlanRequirement 모델의 earned_credit, is_fulfilled 업데이트
+    majors = Major.objects.filter(planmajor__plan=plan)
+    for major in list(majors):
+        mr_r = Requirement.objects.get(major=major, requirement_type=Requirement.MAJOR_REQUIREMENT)
+        me_r = Requirement.objects.get(major=major, requirement_type=Requirement.MAJOR_ELECTIVE)
+        g_r = Requirement.objects.get(major=major, requirement_type=Requirement.GENERAL)
+        ge_r = Requirement.objects.get(major=major, requirement_type=Requirement.GENERAL_ELECTIVE)
+
+        mr_pr = PlanRequirement.objects.get(plan=plan, requirement__major=major,
+                                            requirement__requirement_type=Requirement.MAJOR_REQUIREMENT)
+        me_pr = PlanRequirement.objects.get(plan=plan, requirement__major=major,
+                                            requirement__requirement_type=Requirement.MAJOR_ELECTIVE)
+        g_pr = PlanRequirement.objects.get(plan=plan, requirement__major=major,
+                                           requirement__requirement_type=Requirement.GENERAL)
+        ge_pr = PlanRequirement.objects.get(plan=plan, requirement__major=major,
+                                            requirement__requirement_type=Requirement.GENERAL_ELECTIVE)
+        mr_credit = 0
+        me_credit = 0
+        g_credit = 0
+        ge_credit = 0
+        for semester in list(semesters):
+            mr_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                                 semesterlecture__pivot_major=major,
+                                                 semesterlecture__lecture_type=SemesterLecture.MAJOR_REQUIREMENT)
+            me_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                                 semesterlecture__pivot_major=major,
+                                                 semesterlecture__lecture_type=SemesterLecture.MAJOR_ELECTIVE)
+            g_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                                semesterlecture__pivot_major=major,
+                                                semesterlecture__lecture_type=SemesterLecture.GENERAL)
+            ge_lectures = Lecture.objects.filter(semesterlecture__semester=semester,
+                                                 semesterlecture__pivot_major=major,
+                                                 semesterlecture__lecture_type=SemesterLecture.GENERAL_ELECTIVE)
+            for lecture in list(mr_lectures):
+                mr_credit += lecture.credit
+            for lecture in list(me_lectures):
+                me_credit += lecture.credit
+            for lecture in list(g_lectures):
+                g_credit += lecture.credit
+            for lecture in list(ge_lectures):
+                ge_credit += lecture.credit
+
+        mr_pr.earned_credit = mr_credit
+        mr_pr.save()
+        me_pr.earned_credit = me_credit
+        me_pr.save()
+        g_pr.earned_credit = g_credit
+        g_pr.save()
+        ge_pr.earned_credit = ge_credit
+        ge_pr.save()
+
+        if mr_r.required_credit <= mr_pr.earned_credit:
+            mr_pr.is_fulfilled = True
+            mr_pr.save()
+        if me_r.required_credit <= me_pr.earned_credit:
+            me_pr.is_fulfilled = True
+            me_pr.save()
+        if g_r.required_credit <= g_pr.earned_credit:
+            g_pr.is_fulfilled = True
+            g_pr.save()
+        if ge_r.required_credit <= ge_pr.earned_credit:
+            ge_pr.is_fulfilled = True
+            ge_pr.save()
+
+def cal_lecture_type(lecture, majors, entrance_year):
+    result_major = None
+    result_lecture_type = None
+    for major in list(majors):
+        try:
+            majorlecture = MajorLecture.objects.get(major=major,
+                                                    lecture=lecture,
+                                                    start_year__lte=entrance_year,
+                                                    end_year__gte=entrance_year)
+            if cal_priority_lt(majorlecture.lecture_type) > cal_priority_lt(result_lecture_type):
+                result_major = majorlecture.major
+                result_lecture_type = majorlecture.lecture_type
+        except MajorLecture.DoesNotExist:
+            pass
+    return result_major, result_lecture_type
+
+def cal_priority_lt(lecture_type):
+    switcher = {
+        SemesterLecture.MAJOR_REQUIREMENT: 4,
+        SemesterLecture.MAJOR_ELECTIVE: 3,
+        SemesterLecture.TEACHING: 3,
+        SemesterLecture.GENERAL: 2,
+        SemesterLecture.GENERAL_ELECTIVE: 1
+    }
+    return switcher.get(lecture_type, -1) # -1 is error
+
+def lecture_type_to_int(semesterlecture):
+    switcher = {
+        SemesterLecture.MAJOR_REQUIREMENT : 1,
+        SemesterLecture.MAJOR_ELECTIVE : 2,
+        SemesterLecture.TEACHING : 3,
+        SemesterLecture.GENERAL : 4,
+        SemesterLecture.GENERAL_ELECTIVE : 5
+    }
+    return switcher.get(semesterlecture, -1) # -1 is error
